7/19/2022

+++ src/app.ts
@@ -17,29 +17,26 @@ import ROMFile from './types/romFile.js';
 export default async function main(options: Options) {
   // Find all DAT files and parse them
   // TODO(cemmer): move creation and management of progress bars to each of the modules
-  const datScanProgressBar = new ProgressBar('Scanning for DATs', '⏳', 0);
+  const datScanProgressBar = new ProgressBar('Scanning for DATs', '⏳');
   const dats = await new DATScanner(options, datScanProgressBar).parse();
   if (!dats.length) {
+    await ProgressBar.stop();
     Logger.error('\nNo valid DAT files found!');
     process.exit(1);
   }
-  datScanProgressBar
-    .setSymbol('✅')
-    .setProgressMessage(`${dats.length.toLocaleString()} DAT file${dats.length !== 1 ? 's' : ''} parsed`);
+  await datScanProgressBar.done(`${dats.length.toLocaleString()} DAT${dats.length !== 1 ? 's' : ''} found`);
 
   // Find all ROM files and pre-process them
-  const romScanProgressBar = new ProgressBar('Scanning for ROMs', '⏳', 0);
+  const romScanProgressBar = new ProgressBar('Scanning for ROMs', '⏳');
   const romInputs = await new ROMScanner(options, romScanProgressBar).scan();
-  romScanProgressBar
-    .setSymbol('✅')
-    .setProgressMessage(`${romInputs.length.toLocaleString()} ROM file${romInputs.length !== 1 ? 's' : ''} found`);
+  await romScanProgressBar.done(`${romInputs.length.toLocaleString()} ROM${romInputs.length !== 1 ? 's' : ''} found`);
 
   const datProcessProgressBar = new ProgressBar('Processing DATs', '⚙️', dats.length);
   const datsToWrittenRoms = new Map<DAT, Map<Parent, ROMFile[]>>();
 
   await async.eachLimit(dats, 3, async (dat, callback) => {
     const progressBar = new ProgressBar(dat.getName(), '⏳', dat.getParents().length);
-    datProcessProgressBar.increment();
+    await datProcessProgressBar.increment();
 
     // For each DAT, find all ROM candidates
     const romCandidates = await new CandidateGenerator(progressBar).generate(dat, romInputs);
@@ -53,29 +50,27 @@ export default async function main(options: Options) {
     const parentsWithRomFiles = [...writtenRoms.values()]
       .filter((romFiles) => romFiles.length)
       .length;
-    progressBar
-      .setSymbol('✅')
-      .setProgressMessage(`${parentsWithRomFiles.toLocaleString()} ROM${parentsWithRomFiles !== 1 ? 's' : ''} processed`);
+    await progressBar.done(`${parentsWithRomFiles.toLocaleString()} ROM${parentsWithRomFiles !== 1 ? 's' : ''} processed`);
 
     datsToWrittenRoms.set(dat, writtenRoms);
-    progressBar.delete();
+    if (!parentsWithRomFiles) {
+      await progressBar.delete();
+    }
     callback();
   });
 
-  datProcessProgressBar
-    .setSymbol('✅')
-    .setProgressMessage(`${dats.length.toLocaleString()} DAT${dats.length !== 1 ? 's' : ''} processed`);
+  await datProcessProgressBar.done(`${dats.length.toLocaleString()} DAT${dats.length !== 1 ? 's' : ''} processed`);
 
   // Clean the output directories
   if (options.getClean()) {
-    const cleanerProgressBar = new ProgressBar('Cleaning output', '⏳', 0);
+    const cleanerProgressBar = new ProgressBar('Cleaning output', '⏳');
     const allWrittenRomFiles = [...datsToWrittenRoms.values()]
       .flatMap((parentsToRomFiles) => [...parentsToRomFiles.values()])
       .flatMap((romFiles) => romFiles);
     await new OutputCleaner(options, cleanerProgressBar).clean(allWrittenRomFiles);
   }
 
-  ProgressBar.stop();
+  await ProgressBar.stop();
 
   // Write the output report
   new ReportGenerator(options).write(datsToWrittenRoms);
